requirements.txt
# Web 框架
fastapi>=0.95.0
uvicorn[standard]>=0.21.0

# 数据库
sqlalchemy>=2.0.0
alembic>=1.10.0
# psycopg2-binary>=2.9.5  # PostgreSQL 驱动
pymongo>=4.3.0        

# 安全相关
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6

# 类型验证和配置
pydantic>=1.10.0
email-validator>=1.3.0

# Discord API 集成
requests>=2.28.0
aiohttp>=3.8.4

# 其他工具
python-dotenv>=1.0.0
tenacity>=8.2.0  # 重试机制
httpx>=0.23.0    # 异步 HTTP 客户端

# 可选
# redis>=4.5.0    # 如需缓存
# celery>=5.2.0   # 如需后台任务